refactor: Repeater 패키지 이동, LOTTO_PRICE 상수 클래스 이동
